refactor(step-5):test도 repository, service로 패키지 분리
- User의 public method에 대한 테스트도 생성, 구현 예정
