+++ src/test/java/kr/codesqaud/cafe/model/UserTest.java
@@ -1,0 +1,13 @@
+package kr.codesqaud.cafe.model;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class UserTest {
+
+    @Test
+    void validate() {
+        //TODO : 구현
+    }
+}
